+++ misosoup/main.py
@@ -9,13 +9,15 @@ from collections import defaultdict
 import yaml
 import gurobipy as gp
 
-from reframed import Community, solver_instance
+from reframed import solver_instance
 
 from .library.getters import get_biomass, get_exchange_reactions
 from .library.load import introduce_binary_variables, setup_medium
 from .library.readwrite import load_models, read_compounds
 from .library.solve import minimal_communities, minimal_suppliers
 from .library.validate import validate_solution_dict
+
+from .reframed.layered_community import LayeredCommunity
 
 compute_function = defaultdict(lambda: minimal_suppliers, {"min": minimal_communities})
 
@@ -61,7 +63,7 @@ def main(args):
     media = read_compounds(args.media)
     base_medium = media["base_medium"] if "base_medium" in media.keys() else {}
 
-    community = Community("CoI", models, copy_models=False)
+    community = LayeredCommunity("CoI", models, copy_models=False)
 
     solution = defaultdict(dict)